+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Deprecated
 - Passing `callback` as the first argument when adding a select menu to a component executor or column.
 
+### Fixed
+- [yuyo.modals.modal][] and [yuyo.modals.as_modal][] both now properly support DI for the
+  modal's callback.
+
 ### [1.11.0a1] - 2023-04-02
 ### Added
 - A static timeout implementation.